Add files via upload
i mainly do not jump directly in dataset and code , instead i spend for inspection approx. an hour depending on row numbers. In the data i fixed col_names,dtypes,dropped values which were improper and nulls. Interestingly, i used simple arithmetic to solve nan and wrong value in Item column. So you can read the code . Moreover,it was the best way to fill nulls in Total_spent or others by using * or / .Finally, i filled only with Nan column-Transaction_date.As far as we do not work with stakeholder on the project.